backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
python-dotenv==1.0.0
openai==1.3.0
google-api-python-client==2.108.0
